Ajustes todo n el main
