+++ P10.h
@@ -1,0 +1,5 @@
+int mayor(int num1, int num2);
+int menor(int num1, int num2);
+int generar_aleatorios(int num1, int num2);
+bool pregunta(int a, int b);
+void mensaje(bool tipo);
