+++ P4.c
@@ -1,0 +1,12 @@
+#include <stdio.h>
+#include <math.h>
+#include "P4.h"
+
+int multiplo(int a, int b) {
+  if (a % b == 0) {
+    return 1;
+  } else {
+    return 0;
+  }
+}
+
