+++ P4.h
@@ -1,0 +1,1 @@
+int multiplo(int a, int b);
