+++ P6.h
@@ -1,0 +1,2 @@
+int factorial(int n);
+float serie(int x, int n);
