+++ P7.h
@@ -1,0 +1,1 @@
+int reverse(int n);
