+++ P8.h
@@ -1,0 +1,4 @@
+int mayorP8(int num1, int num2);
+int menorP8(int num1, int num2);
+int generar_aleatoriosP8(int num1, int num2);
+bool preguntaP8(int a, int b);
