+++ P9.h
@@ -1,0 +1,5 @@
+int mayorP9(int num1, int num2);
+int menorP9(int num1, int num2);
+int generar_aleatoriosP9(int num1, int num2);
+bool preguntaP9(int a, int b);
+void mensajeP9(bool tipo);
